pyproject.toml
[tool.poetry]
name = "django-prac-004"
version = "0.1.0"
description = ""
authors = ["noname2048 <sungwook.csw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
psycopg2 = "^2.8.6"
coverage = "^5.5"
codecov = "^2.1.11"
django-storages = "^1.11.1"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
Django = "^3.1.7"
psycopg2 = "^2.8.6"
coverage = "^5.5"
codecov = "^2.1.11"
django-storages = "^1.11.1"
gunicorn = "^20.0.4"
django-debug-toolbar = "^3.2"
black = "^20.8b1"
pylint = "^2.7.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
